fix bug where cov info wasn't being printed for some builds
coverage information was being added only if the build wasn't cached
locally, and only if it wasn't attached before. the former was
undesirable as the coverage data is still there and interesting. the
latter prevented information being displayed for sharded tests

note testing for exact coverage file names seems a bit flimsy, but if
the code changes breaking our assumptions it should be simple enough
to catch.


--
MOS_MIGRATED_REVID=85896250