fix accumulating logic around cached tests
there was an error in the accounting of how many tests were actually
executed vs those that were cached. this was a result of not considering
remotely cached tests when accumulating test results. fix the bug and
add some unit tests to prove out functionality, as well as some
javadoc clarifying functionality.

R=janakr


--
MOS_MIGRATED_REVID=86112421