Squashed two changes:
Prepare strict deps for using JavacPathFileManager.

This lays the groundwork for using a JavacPathFileManager and Jimfs to test
strict deps and the dependency extractor, instead of the ad-hoc in-memory
filesystem implementation in InMemoryJavaFileManager. The current
implementation is deeply impressive, but scary to maintain and entirely
unnecessary now that we have nio-based filemangers Jimfs.

RELNOTES: N/A

----------

Remove cushon from blacklist, as per [].


--
MOS_MIGRATED_REVID=86611345